Move the config-related code to its own file.
